requirements.txt
flask
mysql-connector-python
sentence-transformers
faiss-cpu
python-dotenv
plotly
pandas
